djangoproject/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
